ENG-2369: Fixed std dev calculation in rocksdb metrics.
Summary: Multiplying too big uint64 values was causing an overflow when calculating sum of squares for std_dev calculation and was leading to nan/-nan in metrics JSON output.

Test Plan: Added test for described case.

Reviewers: bogdan, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3485

+++ src/yb/rocksdb/util/histogram.cc
@@ -23,7 +23,7 @@
 
 #include "yb/rocksdb/util/histogram.h"
 
-#include <cassert>
+#include <assert.h>
 #include <math.h>
 #include <stdio.h>
 
@@ -107,7 +107,7 @@ void HistogramImpl::Add(uint64_t value) {
 
   add_to_num(1);
   add_to_sum(value);
-  add_to_sum_squares(value * value);
+  add_to_sum_squares(static_cast<double>(value) * value);
 
   // TODO: maybe we should use int for IndexForValue's return value?
   // There are only 200 or so buckets.
@@ -224,4 +224,4 @@ void HistogramImpl::Data(HistogramData * const data) const {
   data->standard_deviation = StandardDeviation();
 }
 
-} // namespace levedb
+} // namespace rocksdb
